// tiryaq/types/medical.ts
export interface Lesson {
  id: string;
  title: string;
  duration: string;
  type: 'video' | 'resource' | 'quiz';
  free: boolean;
  completed?: boolean;
  description?: string;
  resources?: { id: string; name: string; type: 'pdf' | 'slide' | 'reference'; size: string }[];
}

export interface CourseModule {
  id: string;
  title: string;
  lessons: Lesson[];
}